requirements.txt
pandas
numpy
scipy
matplotlib
seaborn
scikit-learn
pyyaml